Update OOM script
The instructions to deploy ONAP thru OOM tool has been changed, it's
using Helm Charts for performing the deployment. This change reflects
that change.

Given that using Rancher adds an additional layer to the deployment,
it was provided an alternative method to deploy Kubernetes via
kubespray. This tool provides a production ready k8s deployment.

Change-Id: Ied3f5fc9e5c97b4c0f8e623b9d6f3c4f52fc822e
Signed-off-by: Victor Morales <victor.morales@intel.com>
Issue-ID: INT-478